Sample URLs from a pseudo zipf distribution
The random selection of URLS for the posts and
products benchmarks was poor because it's not
realistic and produces a very high cache miss
rate. Real traffic hits some URLs much more than
others. Address this by sampling 5000 URLs using
a pseudo zipf distribution. This allows comparing
the actual caching functionality of the plugins,
and not just their miss behavior.